fix: open in new tab not working
